requirements.txt
feedparser
python-dateutil
praw
python-dotenv
